llm-v1.5 1/requirements_gcp.txt
python-mecab-ko==1.3.7
torch==2.7.0
torchaudio==2.7.0
torchvision==0.22.0
faiss-gpu==1.7.2
PyYAML==6.0.2
pandas==2.2.3
langchain_core==0.3.63
langchain_community==0.3.24
kiwipiepy==0.21.0
konlpy==0.6.0
transformers==4.52.4
ollama==0.5.1
accelerate==1.7.0
FlagEmbedding==1.3.5
rank_bm25==0.2.2
click==8.2.1
numpy==1.26.4